requirements.txt
discord.py==1.7.3
beautifulsoup4==4.10.0
zenrows==0.1.0
aiofiles==0.7.0
aiohttp==3.8.1
markovify==latest_version
Pillow==latest_version
demapi==latest_version